Refactor testing directory architecture
