+++ package.json
@@ -5,7 +5,7 @@
   "main": "index.js",
   "scripts": {
     "dev:server": "ENV=dev ts-node-dev -r tsconfig-paths/register --transpile-only --ignore-watch node_modules src/server.ts",
-    "dev:test": "ENV=test mocha -r ts-node/register -r tsconfig-paths/register --exit src/test/*.test.ts"
+    "dev:test": "ENV=test mocha -r ts-node/register -r tsconfig-paths/register --exit src/test/index.ts"
   },
   "repository": {
     "type": "git",
