+++ src/test/index.ts
@@ -1,0 +1,8 @@
+import 'reflect-metadata';
+import { setup } from 'setup-server';
+before(async () => {
+  await setup();
+});
+
+require('test/login.test.ts');
+require('test/create-user.test.ts');
